Add Datadog log processing and Claude analysis helper
