+++ slack-bot/index.ts
@@ -1,8 +1,12 @@
-import fs from "fs";
-import path from "path";
 import { createLogger } from "@helpers/logger";
 import pkg from "@slack/bolt";
 import dotenv from "dotenv";
+import {
+  askClaude,
+  formatSlackResponse,
+  processDatadogLogs,
+  readIssuesData,
+} from "./helper";
 
 const { App, LogLevel } = pkg;
 
@@ -10,31 +14,6 @@ dotenv.config();
 
 // Initialize logger
 const logger = createLogger();
-
-interface TestFailure {
-  testName: string | null;
-  environment: string | null;
-  geolocation: string | null;
-  timestamp: string | null;
-  workflowUrl: string | null;
-  dashboardUrl: string | null;
-  customLinks: string | null;
-  errorLogs: string[];
-}
-
-interface IssuesData {
-  metadata: {
-    source: string;
-    date: string;
-    totalTestFailures: number;
-    totalLogEntries: number;
-    queryPeriod: {
-      from: string;
-      to: string;
-    };
-  };
-  testFailures: TestFailure[];
-}
 
 // Validate required environment variables
 const requiredEnvVars = ["SLACK_BOT_TOKEN", "SLACK_APP_TOKEN"];
@@ -56,102 +35,62 @@ const app = new App({
   logLevel: LogLevel.ERROR,
 });
 
-// Read test failures from issues.json
-function readTestFailures(): { failures: TestFailure[]; lastUpdate: Date } {
+// Auto-refresh issues.json from Datadog every 10 minutes
+let refreshInterval: NodeJS.Timeout | null = null;
+
+async function refreshIssuesData(): Promise<void> {
   try {
-    const issuesPath = path.join(__dirname, "issues.json");
-    const issuesData = fs.readFileSync(issuesPath, "utf8");
-    const parsedData: IssuesData = JSON.parse(issuesData);
-    return {
-      failures: parsedData.testFailures,
-      lastUpdate: new Date(parsedData.metadata.date),
-    };
+    logger.info("🔄 Refreshing issues data from Datadog...");
+    await processDatadogLogs();
+    logger.info("✅ Issues data refreshed successfully");
   } catch (error) {
     logger.error(
-      `Error reading issues.json: ${error instanceof Error ? error.message : String(error)}`,
+      `❌ Failed to refresh issues data: ${error instanceof Error ? error.message : String(error)}`,
     );
-    return { failures: [], lastUpdate: new Date() };
   }
 }
 
-// Format test failures for Slack display
-function formatTestFailures(failures: TestFailure[], lastUpdate: Date): string {
-  if (failures.length === 0) {
-    return `📊 No test failures found (Last updated: ${lastUpdate.toLocaleTimeString()})`;
-  }
+function startAutoRefresh(): void {
+  // Initial refresh
+  void refreshIssuesData();
 
-  const grouped = failures.reduce<Record<string, TestFailure[]>>(
-    (acc, failure) => {
-      const key = failure.testName || "Unknown";
-      if (!acc[key]) {
-        acc[key] = [];
-      }
-      acc[key].push(failure);
-      return acc;
+  // Set up 10-minute interval (600,000 ms)
+  refreshInterval = setInterval(
+    () => {
+      void refreshIssuesData();
     },
-    {},
+    10 * 60 * 1000,
   );
-
-  let result = `📊 Found ${failures.length} test failures (Last updated: ${lastUpdate.toLocaleTimeString()}):\n\n`;
-
-  for (const [testName, testFailures] of Object.entries(grouped)) {
-    result += `**${testName}** (${testFailures.length} failures)\n`;
-
-    for (const failure of testFailures.slice(0, 3)) {
-      const env = failure.environment || "unknown";
-      const region = failure.geolocation || "unknown";
-      const time = failure.timestamp
-        ? new Date(failure.timestamp).toLocaleTimeString()
-        : "unknown";
-
-      result += `• ${env}/${region} at ${time}\n`;
-      if (failure.errorLogs[0]) {
-        result += `  \`${failure.errorLogs[0].substring(0, 100)}${failure.errorLogs[0].length > 100 ? "..." : ""}\`\n`;
-      }
-    }
-
-    if (testFailures.length > 3) {
-      result += `  ... and ${testFailures.length - 3} more\n`;
-    }
-    result += "\n";
-  }
-
-  return result;
+  logger.info(
+    "🔄 Auto-refresh started: issues.json will update every 10 minutes",
+  );
 }
 
-// Handle messages about test failures
-function handleMessage(message: string): string {
-  const { failures, lastUpdate } = readTestFailures();
-  const lowerMessage = message.toLowerCase();
+function stopAutoRefresh(): void {
+  if (refreshInterval) {
+    clearInterval(refreshInterval);
+    refreshInterval = null;
+    logger.info("🛑 Auto-refresh stopped");
+  }
+}
 
-  // Filter failures based on question
-  if (lowerMessage.includes("browser")) {
-    const browserFailures = failures.filter(
-      (f) =>
-        f.testName?.toLowerCase().includes("browser") ||
-        f.testName?.toLowerCase().includes("playwright"),
-    );
-    return formatTestFailures(browserFailures, lastUpdate);
+// Handle messages using Claude analyzer
+async function handleMessage(message: string): Promise<string> {
+  const rawData = readIssuesData();
+
+  if (!rawData) {
+    return "❌ Could not read issues.json file";
   }
 
-  if (lowerMessage.includes("agent") || lowerMessage.includes("bot")) {
-    const agentFailures = failures.filter(
-      (f) =>
-        f.testName?.toLowerCase().includes("agent") ||
-        f.testName?.toLowerCase().includes("bot"),
+  try {
+    const response = await askClaude(message, rawData);
+    return formatSlackResponse(response);
+  } catch (error) {
+    logger.error(
+      `Error getting Claude response: ${error instanceof Error ? error.message : String(error)}`,
     );
-    return formatTestFailures(agentFailures, lastUpdate);
+    return "❌ Error analyzing issues data";
   }
-
-  if (lowerMessage.includes("group")) {
-    const groupFailures = failures.filter((f) =>
-      f.testName?.toLowerCase().includes("group"),
-    );
-    return formatTestFailures(groupFailures, lastUpdate);
-  }
-
-  // Default: show all failures
-  return formatTestFailures(failures, lastUpdate);
 }
 
 // Respond to @mentions
@@ -163,10 +102,10 @@ app.event<"app_mention">("app_mention", async ({ event, say, client }) => {
 
     logger.info(`📨 RECEIVED MENTION - Channel: ${channel}, User: ${userId}`);
 
-    const thinkingMessage = `<@${userId}> 🤔 Checking issues.json...`;
+    const thinkingMessage = `<@${userId}> 🤔 *Thinking...*`;
     const thinkingResponse = await say(thinkingMessage);
 
-    const botResponse = handleMessage(message);
+    const botResponse = await handleMessage(message);
     const finalResponse = `<@${userId}> ${botResponse}`;
 
     // Replace the thinking message with the final response
@@ -183,7 +122,46 @@ app.event<"app_mention">("app_mention", async ({ event, say, client }) => {
     logger.error(
       `Error processing app mention: ${error instanceof Error ? error.message : String(error)}`,
     );
-    const errorResponse = `<@${event.user}> Sorry, I encountered an error reading issues.json.`;
+    const errorResponse = `<@${event.user}> 🚨 *Error:* Sorry, I encountered an error analyzing the data.`;
+    await say(errorResponse);
+  }
+});
+
+// Handle /qa slash command
+app.command("/qa", async ({ command, ack, say, client }) => {
+  try {
+    // Acknowledge the command request
+    await ack();
+
+    const messageText = command.text || "What are the current issues?";
+    const userId = command.user_id;
+    const channel = command.channel_id;
+
+    logger.info(
+      `🔸 RECEIVED SLASH COMMAND - Channel: ${channel}, User: ${userId}`,
+    );
+
+    const thinkingMessage = `<@${userId}> 🤔 *Thinking...*`;
+    const thinkingResponse = await say(thinkingMessage);
+
+    const botResponse = await handleMessage(messageText);
+    const finalResponse = `<@${userId}> ${botResponse}`;
+
+    // Replace the thinking message with the final response
+    if (thinkingResponse && thinkingResponse.ts) {
+      await client.chat.update({
+        channel: channel,
+        ts: thinkingResponse.ts,
+        text: finalResponse,
+      });
+    } else {
+      await say(finalResponse);
+    }
+  } catch (error: unknown) {
+    logger.error(
+      `Error processing slash command: ${error instanceof Error ? error.message : String(error)}`,
+    );
+    const errorResponse = `<@${command.user_id}> 🚨 *Error:* Sorry, I encountered an error analyzing the data.`;
     await say(errorResponse);
   }
 });
@@ -212,10 +190,10 @@ app.message(async ({ message, say, client }) => {
 
       logger.info(`📨 RECEIVED DM - User: ${userId}`);
 
-      const thinkingMessage = "🤔 Checking issues.json...";
+      const thinkingMessage = "🤔 *Thinking...*";
       const thinkingResponse = await say(thinkingMessage);
 
-      const botResponse = handleMessage(messageText);
+      const botResponse = await handleMessage(messageText);
 
       // Replace the thinking message with the final response
       if (thinkingResponse && thinkingResponse.ts) {
@@ -232,7 +210,7 @@ app.message(async ({ message, say, client }) => {
     logger.error(
       `Error processing direct message: ${error instanceof Error ? error.message : String(error)}`,
     );
-    await say("Sorry, I encountered an error reading issues.json.");
+    await say("🚨 *Error:* Sorry, I encountered an error analyzing the data.");
   }
 });
 
@@ -241,6 +219,9 @@ void (async () => {
   try {
     await app.start();
     logger.info("🚀 Slack bot started successfully");
+
+    // Start auto-refresh for issues data
+    startAutoRefresh();
   } catch (error: unknown) {
     logger.error(
       `Failed to start application: ${error instanceof Error ? error.message : String(error)}`,
@@ -252,10 +233,12 @@ void (async () => {
 // Graceful shutdown
 process.on("SIGTERM", () => {
   logger.info("🛑 Shutting down gracefully");
+  stopAutoRefresh();
   process.exit(0);
 });
 
 process.on("SIGINT", () => {
   logger.info("🛑 Shutting down gracefully");
+  stopAutoRefresh();
   process.exit(0);
 });
